refactor: splitting useReleaseTime from ReleaseTime component
